+++ src/modules/species/dtos/responses/pokemon-response.dto.ts
