+++ src/modules/species/enums/translation-type.enum.ts
