+++ src/modules/species/interfaces/pokemon-species-api-response.ts
