+++ src/modules/species/interfaces/translated-text-api-response.ts
