+++ src/modules/species/species.service.ts
@@ -5,8 +5,8 @@ import { firstValueFrom } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { ThrowError } from 'src/utils/error-handler';
 import { normalizeTextFormatting } from 'src/utils/string-handler';
-import { PokemonResponseDto } from './dto/responses/pokemon-response.dto';
-import { PokemonTranslatedResponseDto } from './dto/responses/pokemon-translated-response.dto';
+import { PokemonResponseDto } from './dtos/responses/pokemon-response.dto';
+import { PokemonTranslatedResponseDto } from './dtos/responses/pokemon-translated-response.dto';
 import { TranslationTypeEnum } from './enums/translation-type.enum';
 import { PokemonSpeciesApiResponse } from './interfaces/pokemon-species-api-response';
 import { TranslatedTextApiResponse } from './interfaces/translated-text-api-response';
